Move some files, and move commands into cogs
