+++ src/Codeforces/Funcs.py
